refactor(BettingOdds, BettingResults): 적절한 패키지로 이동
