+++ src/main/java/model/result/BettingResults.java
@@ -1,4 +1,4 @@
-package model.bet;
+package model.result;
 
 import java.util.Map;
