+++ src/test/java/model/BlackjackGameTest.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import model.bet.BettingResults;
+import model.result.BettingResults;
 import model.bet.ParticipantsBet;
 import model.card.Card;
 import model.card.CardNumber;
@@ -19,7 +19,6 @@ import model.cards.DealerCards;
 import model.cards.ParticipantsCards;
 import model.cards.PlayerCards;
 import model.deck.Deck;
-import model.result.GameResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
