+++ src/test/java/model/bet/BettingResultsTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Map;
+import model.result.BettingResults;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
